refactor: msw participatedMentoring 파일 경로 변경 #894
