+++ frontend/src/pages/participatedMentoring/MentoringItem/MentoringItem.stories.tsx
@@ -1,7 +1,7 @@
 import { MemoryRouter } from 'react-router-dom';
 
 import { PAGE_URL } from '../../../common/constants/url';
-import { PARTICIPATED_MENTORING_LIST } from '../../../common/mock/participatedMentoring/data';
+import { PARTICIPATED_MENTORING_LIST } from '../mock/data';
 
 import MentoringItem from './MentoringItem';
