+++ frontend/src/pages/participatedMentoring/mock/handler.ts
@@ -1,6 +1,6 @@
 import { http, HttpResponse } from 'msw';
 
-import { API_ENDPOINTS } from '../../constants/apiEndpoints';
+import { API_ENDPOINTS } from '../../../common/constants/apiEndpoints';
 
 import { PARTICIPATED_MENTORING_LIST } from './data';
